Update Duke to allow different types of UI
