+++ app/src/main/java/duke/GraphicalUi.java
@@ -1,0 +1,19 @@
+package duke;
+
+public class GraphicalUi extends Ui {
+    /**
+     * Creates a new GUI handler.
+     */
+    public GraphicalUi() {
+    }
+
+    @Override
+    public String getLine() {
+        return "";
+    }
+
+    @Override
+    public void respond(String response) {
+        //
+    }
+}
