+++ app/src/main/java/duke/Ui.java
@@ -1,39 +1,21 @@
 package duke;
 
-import java.util.Scanner;
-
-public class Ui {
-    private Scanner sc;
-
+public abstract class Ui {
     /**
      * Creates a new UI handler.
      */
     public Ui() {
-        sc = new Scanner(System.in);
     }
 
     /**
      * Gets a line of user input.
      * @return The user input.
      */
-    public String getLine() {
-        return sc.nextLine();
-    }
-
-    /**
-     * Displays a divider.
-     */
-    public void showDivider() {
-        System.out.println("    __________________________________________________");
-    }
+    public abstract String getLine();
 
     /**
      * Display some output to the user.
      * @param response The output to be shown.
      */
-    public void respond(String response) {
-        showDivider();
-        System.out.println("    " + response.replace("\n", "\n    "));
-        showDivider();
-    }
+    public abstract void respond(String response);
 }
